a few fixes made during initial testing

+++ CoggingCompTraj.h
@@ -47,7 +47,7 @@ private:
 
   bool first;
   int sample_count;
-  int last_position;
+  double last_position;
   std::vector<double> fixed_pos;
   double sample_time;
   double done_time;
